Refactor
- Moved all const to constants.go
- Moved sqlEvent to event.go
- Added utils.go
